Split targets.py along its own seams
targets.py held four separable things: the regime thresholds and their verdict,
the Benchmark abstraction, and the two benchmark families. Each is now its own
module:

  regime.py         E-L*'s operating regime: thresholds + report() for one DFA
  benchmark.py      the Benchmark dataclass + the upstream -> automata-lib port
  our_targets.py    family 1, this repo's oracles
  capal_targets.py  family 2, CAPAL's shipped .taf dataset

Benchmark.accept_rate and .class_preserving_frac existed only to feed
regime_report, so they are absorbed into regime.report, which now takes the
automata-lib DFA directly. Benchmark.regime_report stays as the caller-facing
entry point.

Verified behaviour-preserving: regime reports for all 33 targets are unchanged
except `uncovered_states` on regex_alt_111_or_000_3sym, which varies across
processes on unchanged code too -- build_regex_dfa numbers its states via
hash-ordered iteration, so runs produce isomorphic DFAs with different labels.
